pyproject.toml
[tool.poetry]
name = "blogrest"
version = "0.1.0"
description = ""
authors = ["seth mcfeeters <seth.mcfeeters@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
black = "^20.8b1"
flake8 = "^3.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
